+++ packages/orq-rc/src/models/operations/createidentity.ts
@@ -139,7 +139,7 @@ export const CreateIdentityResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-10T09:05:30.137Z",
+    "2026-02-10T11:08:07.729Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {