+++ packages/orq-rc/RELEASES.md
@@ -6579,3 +6579,13 @@ Based on:
 - [typescript v4.3.0-rc.43] packages/orq-rc
 ### Releases
 - [NPM v4.3.0-rc.43] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.43 - packages/orq-rc
+
+## 2026-02-10 11:13:55
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.709.3 (2.814.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.3.0-rc.45] packages/orq-rc
+### Releases
+- [NPM v4.3.0-rc.45] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.45 - packages/orq-rc